requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aiofiles==23.2.1
pydantic[email]==2.7.4
psycopg2-binary==2.9.9
supabase==2.7.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
# Dev / testing
pytest==8.2.2
httpx==0.27.0
